+++ HaloEzAPI/Abstraction/Enum/AccessControl.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HaloEzAPI.Abstraction.Enum
+{
+    public enum AccessControl
+    {
+        Listed = 0,
+        Unlisted = 1,
+        Unknown = 2,
+    }
+}