+++ HaloEzAPI/Model/Request/Order.cs
@@ -1,0 +1,8 @@
+﻿namespace HaloEzAPI.Model.Request
+{
+    public enum Order
+    {
+        Asc,
+        Desc,
+    }
+}