+++ HaloEzAPI/Model/Request/Sort.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HaloEzAPI.Model.Request
+{
+    public enum Sort
+    {
+        Name,
+        Description,
+        Accessibility,
+        Created,
+        Modified,
+        BookmarkCount,
+    }
+}