+++ HaloEzAPI/Model/Response/UGC/TimeUtc.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace HaloEzAPI.Model.Response.UGC
+{
+    public class TimeUtc
+    {
+        [JsonConverter(typeof(IsoDateTimeConverter))]
+        public DateTime ISO8601Date { get; set; }
+    }
+}