+++ HaloEzAPI/Model/Response/UGC/UGCGameVariant.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HaloEzAPI.Model.Response.UGC
+{
+    public class UGCGameVariant : UGCBase
+    {
+        public int BaseEngineGameType { get; set; }
+        public int ScoreToWin { get; set; }
+        public int NumberOfLives { get; set; }
+        public int MatchDurationInSeconds { get; set; }
+    }
+}