+++ HaloEzAPI/Model/Response/UGC/UGCStats.cs
@@ -1,0 +1,8 @@
+﻿namespace HaloEzAPI.Model.Response.UGC
+{
+    public class UGCStats
+    {
+        public int BookmarkCount { get; set; }
+        public bool HasCallerBookmarked { get; set; }
+    }
+}