+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changlog
 ---------------
+### 1.3.0 **Release Date: 17/11/2016**
+- Added UGC Endpoints
+- Added in Player Leaderboard Endpoint
+- Additional types
 
-### 1.2.6
+### 1.2.6 **Release Date: N/A**
 - Added in `Infection` to `Medal.cs`
 
-### 1.2.5
+### 1.2.5 **Release Date: N/A**
 - Fix to `Guid` translation using JacobSnyders branch (manual)
 - Adjustmnet to SeasonId
 - Additional tests